Moving wrapper objects into their own package
These are just convenience objects for Gson parsing of API calls that return an object with a list instead of just the plain list.
