+++ src/main/java/edu/ksu/canvas/impl/QuizSubmissionQuestionImpl.java
@@ -16,8 +16,8 @@ import edu.ksu.canvas.interfaces.QuizSubmissionQuestionReader;
 import edu.ksu.canvas.interfaces.QuizSubmissionQuestionWriter;
 import edu.ksu.canvas.model.assignment.QuizAnswer;
 import edu.ksu.canvas.model.assignment.QuizSubmissionQuestion;
-import edu.ksu.canvas.model.assignment.QuizSubmissionQuestionWrapper;
-import edu.ksu.canvas.model.assignment.QuizSubmissionWrapper;
+import edu.ksu.canvas.model.wrapper.QuizSubmissionQuestionWrapper;
+import edu.ksu.canvas.model.wrapper.QuizSubmissionWrapper;
 import edu.ksu.canvas.net.Response;
 import edu.ksu.canvas.net.RestClient;
 import edu.ksu.canvas.oauth.OauthToken;
