+++ src/main/java/edu/ksu/canvas/model/wrapper/QuizSubmissionQuestionWrapper.java
@@ -1,6 +1,8 @@
-package edu.ksu.canvas.model.assignment;
+package edu.ksu.canvas.model.wrapper;
 
 import java.util.List;
+
+import edu.ksu.canvas.model.assignment.QuizSubmissionQuestion;
 
 /**
  * Wrapper class made necessary because canvas returns an object that
