+++ src/main/java/edu/ksu/canvas/model/wrapper/QuizSubmissionWrapper.java
@@ -1,6 +1,8 @@
-package edu.ksu.canvas.model.assignment;
+package edu.ksu.canvas.model.wrapper;
 
 import java.util.List;
+
+import edu.ksu.canvas.model.assignment.QuizSubmission;
 
 /**
  * Wrapper class made necessary because canvas returns an object that
