+++ src/main/java/edu/ksu/canvas/model/wrapper/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Wrapper objects for when the Canvas API calls
+ * <p>
+ * Some API calls return an object with a list inside of it
+ * instead of just the list. The classes in here are just
+ * objects to hold lists for easy parsing via Gson. They
+ * should only be used internally to the API library.
+ */
+package edu.ksu.canvas.model.wrapper;
